+++ drupal/acquia/list.sh
@@ -20,7 +20,7 @@ echo "List data for Howard D8 sites."
 
 # Choose what to list.
 echo "What would you like to list?"
-LIST_TYPES=( "users" "webforms" "newsfeeds" "magazinefeeds" )
+LIST_TYPES=( "users" "webforms" "webform_emails" "newsfeeds" "magazinefeeds" )
 select LIST_TYPE in "${LIST_TYPES[@]}"; do
   if [[ -z "$LIST_TYPE" ]]; then
     printf '"%s" is not a valid choice\n' "$REPLY" >&2
@@ -30,10 +30,11 @@ select LIST_TYPE in "${LIST_TYPES[@]}"; do
 done
 
 case $LIST_TYPE in
-  "users")         LABEL="user lists";         REMOTE_SCRIPT="hal_user_list.sh" ;;
-  "webforms")      LABEL="webform lists";       REMOTE_SCRIPT="hal_webform_list.sh" ;;
-  "newsfeeds")     LABEL="news feed lists";     REMOTE_SCRIPT="hal_newsfeed_list.sh" ;;
-  "magazinefeeds") LABEL="magazine feed lists"; REMOTE_SCRIPT="hal_magazinefeed_list.sh" ;;
+  "users")         LABEL="user lists";               REMOTE_SCRIPT="hal_user_list.sh" ;;
+  "webforms")      LABEL="webform lists";             REMOTE_SCRIPT="hal_webform_list.sh" ;;
+  "webform_emails") LABEL="webform email handler lists"; REMOTE_SCRIPT="hal_webform_email_list.sh" ;;
+  "newsfeeds")     LABEL="news feed lists";           REMOTE_SCRIPT="hal_newsfeed_list.sh" ;;
+  "magazinefeeds") LABEL="magazine feed lists";       REMOTE_SCRIPT="hal_magazinefeed_list.sh" ;;
 esac
 
 DIR="${BASH_SOURCE%/*}"